+++ backend/src/main/java/edu/sjsu/directexchange/service/UserService.java
@@ -4,7 +4,7 @@ import edu.sjsu.directexchange.model.User;
 
 public interface UserService {
 
-  public User getUserById(Long id);
+  public User getUserById(int id);
 
   public void createUser(User user);
